actions divided in two files: actions.js (users) & petsActios.js
